[#178] fix: 페이지 내 기능 변경 및 오류 수정
