+++ src/main/resources/static/js/map.js
@@ -1,16 +1,3 @@
-document.addEventListener("DOMContentLoaded", function () {
-    const academyFeature = document.getElementById("academy-feature");
-
-    if (academyFeature) {
-        academyFeature.addEventListener("click", function () {
-            const url = `/map`;
-            const name = 'popup-map'
-            const options = 'top=10, left=10, width=800, height=600, status=no, menubar=no, toolbar=no, resizable=no, scrollbars=no, fullscreen=no';
-            window.open(url, name, options);
-        });
-    }
-});
-
 const loadingBar = document.getElementById('loading-bar');
 loadingBar.style.display = 'block';
