+++ src/main/resources/static/js/scriptDetail.js
@@ -27,7 +27,6 @@ document.querySelector('.close-btn').addEventListener('click', function() {
     document.getElementById('tooltip-modal').style.display = 'none';
 });
 
-// 팝업 외부 클릭 시 팝업 닫기
 window.addEventListener('click', function(event) {
     const modal = document.getElementById('tooltip-modal');
     if (event.target === modal) {
