+++ src/main/resources/static/js/userMe.js
@@ -55,11 +55,11 @@ function showLoggedInNav(userData) {
         `
     }
     navButtons.innerHTML = `
+        <a href="/" class="nav-button" >홈</a>
+        <a href="/boards" class="nav-button">게시판</a>
+        <a href="/scripts-list" class="nav-button">스피치 분석</a>
         <a href="/page/me" class="nav-button">마이페이지</a>
         <a onclick=logout() class="nav-button" id="logout-button">로그아웃</a>
-        <a href="/" class="nav-button" >홈</a>
-        <a href="/script-list" class="nav-button">스피치 분석</a>
-        <a href="/boards" class="nav-button">게시판</a>
     `;
 }
 
@@ -87,8 +87,8 @@ function myPage(userData){
 function showLoggedOutNav() {
     const navButtons = document.getElementById('nav-buttons');
     navButtons.innerHTML = `
+        <a href="/" class="nav-button" >홈</a>
         <a href="/login" class="nav-button">로그인</a>
-        <a href="/boards" class="nav-button">게시판</a>
     `;
 }
 function logout(){
