+++ src/main/resources/templates/home.html
@@ -9,8 +9,6 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
     <link rel="stylesheet" href="/css/land-style.css">
 </head>
-<script src="/scriptPage/js/userMe.js"></script>
-<script src="/scriptPage/js/addTokenSession.js"></script>
 <body>
 <header>
     <div id="nav-buttons">
@@ -24,30 +22,30 @@
     <div class="container">
         <div class="row text-center">
             <div class="col-md-4">
-                <div class="feature-box p-4 mb-4">
+                <div id="board" class="feature-box p-4 mb-4">
                     <div class="feature-icon mb-3"><i class="fas fa-microphone"></i></div>
-                    <h3>Record Scripts</h3>
-                    <p>원하는 대본을 작성해보세요.</p>
+                    <h3>🙋‍ 커뮤니케이션</h3>
+                    <p>사람들과 스피치에 대한 얘기를 나눠보세요.</p>
                 </div>
             </div>
             <div class="col-md-4">
-                <div class="feature-box p-4 mb-4">
+                <div id="analyze-record" class="feature-box p-4 mb-4">
                     <div class="feature-icon mb-3"><i class="fas fa-brain"></i></div>
-                    <h3>AI Processing</h3>
-                    <p>AI를 통해 당신의 점수를 확인하세요.</p>
+                    <h3>🤖 스피치 분석</h3>
+                    <p>대본에 대한 녹음 파일로 AI를 통해 당신의 점수를 확인하세요.</p>
                 </div>
             </div>
-            <div class="col-md-4">
+            <!--<div class="col-md-4">
                 <div class="feature-box p-4 mb-4">
                     <div class="feature-icon mb-3"><i class="fas fa-cloud-upload-alt"></i></div>
                     <h3>Cloud Storage</h3>
                     <p>점수를 저장하여 발전한 모습을 스스로 확인해 보세요.</p>
                 </div>
-            </div>
+            </div>-->
             <div class="col-md-4">
                 <div id="academy-feature" class="feature-box p-4 mb-4">
                     <div class="feature-icon mb-3"><i class="fas fa-academy"></i></div>
-                    <h3>Speech Academy</h3>
+                    <h3>🧭 내 주변 스피치 학원</h3>
                     <p>회원님 주변의 스피치 학원을 찾아드릴게요.</p>
                 </div>
             </div>
@@ -67,6 +65,8 @@
         <p>&copy; 2024 AI Script Recorder & Processor. All rights reserved.</p>
     </div>
 </footer>
-<script src="/js/map.js"></script>
+<script src="/js/home.js"></script>
+<script src="/js/userMe.js"></script>
+<script src="/js/addTokenSession.js"></script>
 </body>
 </html>
